feat(Move type declarations to types file): The move to a central types file (.d.ts)
Required removing the babel types import in the declaration file as well as
removing many import statements.

re #55
