+++ package.json
@@ -36,6 +36,7 @@
   "homepage": "https://teleporthq.io/",
   "devDependencies": {
     "@fabiospampinato/bump": "^1.0.6",
+    "@types/babel-types": "^7.0.6",
     "@types/jest": "^24.0.6",
     "@types/node": "^11.9.4",
     "@types/rimraf": "^2.0.2",
