+++ src/core/builder/generators/html-to-string.ts
@@ -6,7 +6,6 @@ import parserPostCSS from 'prettier/parser-postcss'
 import toHTML from 'hast-util-to-html'
 
 import { PRETTIER_CONFIG } from '../../../shared/constants'
-import { CodeGeneratorFunction, HastNode } from '../../../shared/types'
 
 export const generator: CodeGeneratorFunction<HastNode> = (htmlObject) => {
   const unformatedString = toHTML(htmlObject)
