+++ src/index.ts
@@ -1,7 +1,4 @@
 import * as UIDLValidators from './uidl-definitions/validators'
-import * as UIDLTypes from './uidl-definitions/types'
-
-import * as GeneratorTypes from './shared/types'
 
 // Different project flavor creators
 export { default as createReactBasicProject } from './project-generators/react-basic'
@@ -15,4 +12,4 @@ export {
 } from './component-generators/react/react-component'
 export { default as createVueComponentGenerator } from './component-generators/vue/vue-component'
 
-export { UIDLValidators, UIDLTypes, GeneratorTypes }
+export { UIDLValidators }
