+++ src/plugins/teleport-plugin-react-css-modules/index.ts
@@ -1,6 +1,5 @@
 import * as t from '@babel/types'
 import { ParsedASTNode } from '../../shared/utils/ast-js-utils'
-import { ComponentPlugin, ComponentPluginFactory } from '../../shared/types'
 import { cammelCaseToDashCase, stringToCamelCase } from '../../shared/utils/string-utils'
 import { addJSXTagStyles, addDynamicAttributeOnTag } from '../../shared/utils/ast-jsx-utils'
 import {
