+++ src/plugins/teleport-plugin-react-proptypes/index.ts
@@ -1,4 +1,3 @@
-import { ComponentPlugin, ComponentPluginFactory } from '../../shared/types'
 import { buildDefaultPropsAst, buildTypesOfPropsAst } from './utils'
 
 interface ReactJSPropTypesConfig {
