+++ src/plugins/teleport-plugin-react-proptypes/utils.ts
@@ -1,5 +1,4 @@
 import * as types from '@babel/types'
-import { PropDefinition } from '../../uidl-definitions/types'
 import { objectToObjectExpression, ParsedASTNode } from '../../shared/utils/ast-js-utils'
 
 export const buildDefaultPropsAst = (
